+++ engines/__init__.py
@@ -1,0 +1,10 @@
+"""
+✨ Thank you for using our Automata Theory Toolkit! ✨
+
+This implementation provides a solid foundation for:
+• Understanding finite automata concepts
+• Learning computational theory
+• Implementing language recognition algorithms
+
+🚀 Built with passion for Computer Science education! 🚀
+"""
